fix blendable datasets + model parallel
